Changed : create a seperate file for v2 parser. Changed : other small improvements : use io.Copy, scan lines on our own, reduce regexp lookups by elimitating index search for subindexes Added : record with predefined size to reduce append(..) calls
